+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @flowglad/react
 
+## 0.10.16
+
+### Patch Changes
+
+- Add getProduct and getPrice to SDK, support activate_subscription checkout sessions
+- Updated dependencies
+  - @flowglad/shared@0.10.16
+
 ## 0.10.15
 
 ### Patch Changes